day_14: Z3 magic!  day_13: not cheating this time :P
Well, day 14 was interesting. Figuring out how to define the problem in
Z3 was difficult. I'm sure there are better ways to describe it, or
possible some data types or tactics that could simplify it. A way to
have integer division round up would be useful, maybe.

Day 13 part 2 was beaten by watching the ball's height and making sure
that we couldn't lose the game by placing it at the top of the screen
when it hit the bottom. After some suggestions on how to actually beat
the game without cheating it seems the solution was rather simple!