days 14,15,16: add solutions
Days 14 and 15 probably need some polishing, but the code is there to
solve the problem. I was very, very lazy on day 15 and it requires
watching part 1 in order to copy-paste the maze output for part2. Day 15
part 2 also has an off-by-one error, not sure if that's due to part 1.

Day 16 is more polished. Part 1 was a straightforward implementation of
the described algorithm and matching the first example was sufficient to
pass. I noticed in part 2 how the pattern carried to large offsets meant
some optimizations were possible.